ユーザー情報APIを統合
